refactor: Notification settings category
