+++ airsync-mac/Localization/en.json
@@ -19,6 +19,7 @@
   "settings.tab": "Settings",
   "settings.myMac": "My Mac",
   "settings.sync": "Sync",
+  "settings.notifications": "Notifications & Alerts",
   "settings.mirroring": "Mirroring",
   "settings.menubar": "Menubar",
   "settings.appearance": "Appearance & more",
