+++ airsync-mac/Screens/Settings/SettingsView.swift
@@ -10,6 +10,8 @@ struct SettingsView: View {
                 MyMacSettingsView()
             case .sync:
                 SyncSettingsView()
+            case .notifications:
+                NotificationsSettingsView()
             case .mirroring:
                 MirroringSettingsView()
             case .quickShare:
